+++ fuel-derive/src/serialize.rs
@@ -1,7 +1,7 @@
 use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
 
-use crate::attribute::{
+use crate::canonical_attribute::{
     should_skip_field_binding,
     StructAttrs,
 };