+++ fuel-tx/src/transaction/types/output.rs
@@ -1,8 +1,8 @@
 use fuel_crypto::Hasher;
 use fuel_types::{
     canonical::{
-        Deserialize,
-        Serialize,
+        self,
+        Serialize as _,
     },
     Address,
     AssetId,
@@ -11,8 +11,6 @@ use fuel_types::{
     Nonce,
     Word,
 };
-
-use core::mem;
 
 mod consts;
 pub mod contract;
@@ -23,25 +21,34 @@ pub use repr::OutputRepr;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, strum_macros::EnumCount)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Deserialize, Serialize)]
+#[cfg_attr(feature = "da-compression", derive(fuel_compression::Compact))]
+#[derive(canonical::Deserialize, canonical::Serialize)]
 pub enum Output {
     Coin {
+        #[cfg_attr(feature = "da-compression", da_compress(registry = "Address"))]
         to: Address,
         amount: Word,
+        #[cfg_attr(feature = "da-compression", da_compress(registry = "AssetId"))]
         asset_id: AssetId,
     },
 
     Contract(Contract),
 
     Change {
+        #[cfg_attr(feature = "da-compression", da_compress(registry = "Address"))]
         to: Address,
+        #[cfg_attr(feature = "da-compression", da_compress(skip))]
         amount: Word,
+        #[cfg_attr(feature = "da-compression", da_compress(registry = "AssetId"))]
         asset_id: AssetId,
     },
 
     Variable {
+        #[cfg_attr(feature = "da-compression", da_compress(skip))]
         to: Address,
+        #[cfg_attr(feature = "da-compression", da_compress(skip))]
         amount: Word,
+        #[cfg_attr(feature = "da-compression", da_compress(skip))]
         asset_id: AssetId,
     },
 
@@ -198,7 +205,7 @@ impl Output {
             Output::Contract(contract) => contract.prepare_sign(),
 
             Output::Change { amount, .. } => {
-                mem::take(amount);
+                *amount = 0;
             }
 
             Output::Variable {
@@ -207,9 +214,9 @@ impl Output {
                 asset_id,
                 ..
             } => {
-                mem::take(to);
-                mem::take(amount);
-                mem::take(asset_id);
+                *to = Address::default();
+                *amount = 0;
+                *asset_id = AssetId::default();
             }
 
             _ => (),
@@ -220,7 +227,7 @@ impl Output {
     pub fn prepare_init_script(&mut self) {
         match self {
             Output::Change { amount, .. } => {
-                mem::take(amount);
+                *amount = 0;
             }
 
             Output::Variable {
@@ -228,9 +235,9 @@ impl Output {
                 amount,
                 asset_id,
             } => {
-                mem::take(to);
-                mem::take(amount);
-                mem::take(asset_id);
+                *to = Address::default();
+                *amount = 0;
+                *asset_id = AssetId::default();
             }
 
             _ => (),